Added treepunching and thieving skills
Added treepunching and updated the onBlockBreakEvent to generalize and
make checks for multiple skills.

Created new mapping on EntityId's, which will be used for methods that
need to look up an ID that represents the type on entity.

Added thieving skill. It contains a mapping of EntityIds to thieving data.
It contains level requirements, experience, probability of  success and a
table of items and probabilities.

Fixed the ItemManaHeal wo work on right click, not when right clicking a
block.