New Thieving Mechanics
Made it so you have to wait a specified number of ticks to attempt stealing again from an NPC. The NPC also has an alert level, which  slowly goes up each time you successfully steal, or
gets really high upon failing. The higher the alert level of the NPC, the lower the probability of stealing is. As time passes, the alert level of the NPC decreases